pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
sphinx = "^8.2.3"
pytest = "^8.3.5"
pytest-cov = "^6.0.0"
httpx = "^0.28.1"

[project]
name = "goit-pythonweb-hw-12"
version = "0.1.0"
description = ""
authors = [
    {name = "Serhii Tarasenko",email = "stdev33@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "annotated-types (==0.7.0)",
    "anyio (==4.8.0)",
    "click (==8.1.8)",
    "dnspython (==2.7.0)",
    "email-validator (==2.2.0)",
    "fastapi (==0.115.11)",
    "h11 (==0.14.0)",
    "idna (==3.10)",
    "psycopg2-binary (==2.9.10)",
    "pydantic (==2.10.6)",
    "pydantic-core (==2.27.2)",
    "python-dotenv (==1.0.1)",
    "sniffio (==1.3.1)",
    "sqlalchemy (==2.0.39)",
    "starlette (==0.46.1)",
    "typing-extensions (==4.12.2)",
    "uvicorn (==0.34.0)",
    "passlib (==1.7.4)",
    "bcrypt (== 4.1.2)",
    "pyjwt (== 2.8.0)",
    "black (>=25.1.0,<26.0.0)",
    "alembic (>=1.15.1,<2.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "sendgrid (>=6.11.0,<7.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "cloudinary (>=1.43.0,<2.0.0)",
    "redis (>=5.2.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
